requirements.txt
torch>=2.0.0
torchvision>=0.15.0
open_clip_torch>=2.20.0
transformers>=4.30.0
pycocotools>=2.0.6
Pillow>=9.0.0
tqdm>=4.60.0
wandb>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0
spacy>=3.5.0
datasets>=2.14.0
matplotlib>=3.7.0
